pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "lexicore-engine"
version = "2.0.0"
description = "Blazing-fast offline dictionary & vocabulary engine"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
authors = [
    {name = "yuki-20"}
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "metaphone>=0.6",
    "python-Levenshtein>=0.25.0",
    "gtts>=2.5.0",
    "aiohttp>=3.10.0",
    "Pillow>=10.4.0",
    "pytesseract>=0.3.13",
    "vosk>=0.3.45",
    "PyAudio>=0.2.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["engine/tests"]
asyncio_mode = "auto"
